requirements.txt
python-dotenv
openai
pyupbit
python-telegram-bot
asyncio